Style: 채팅 레이아웃 v3 적용
